Version 2.0
This is a major revision of winmanager with lots of new features, improved
stability, etc.

IMPORTANT: you need to download and install the latest version of
bufexplorer (vimscript #42) to make this version of winmanager work.

1. One of the major new features of this version is (hopefully, touch wood,
etc) its stability. It has undergone extensive QA by Xiangjiang Ma (and
other users of the 2.0 beta version have also been using it for a couple of
weeks and have reported a few bugs). A big thanks to him (and others) for
all the valuable help. Infact a lot of the new features and the
implementation methods owe their style/origin to his ideas.

2. This version of winmanager no longer ships with a duplicate of
bufexplorer.vim. This has many advantages such as a reduction in script
bloat and in general promotes the idea of modularity. The latest version of
bufexplorer.vim (see above) provides hooks so that winmanager can
communicate with it. (thanks to Jeff Lanzarotta for making the necessary
changes).

3. Becuase of some new features of bufexplorer.vim, winmanager is able to
offer better dynamic behavior. For example, when you open up winmanager
when your editing only a single buffer, bufexplorer will not be shown. As
soon as at least 2 buffers are opened, bufexplorer is automatically opened.

4. This version provides an easy way to customize the window layout of
winmanager using non-default values of the global variable,
g:winManagerWindowLayout. (see ":he winmanager-customizing" for details).

5. This version also provides a way to use winmanager.vim to "edit"
directories instead of the standard exploerer.vim. To do this, you will
need to disable the standard explorer.vim (either by renaming it to
explorer.txt or putting ":let loaded_explorer = 1" in your .vimrc). You
then need to put
   :let g:defaultExplorer = 0
in your .vimrc to let winmanager know that its supposed to take over.

6. I have tried to set-up a framework wherby it is very easy to add other
IDE type plugins to winmanager. By this I mean that a plugin author
could, with the minimum of effort make a plugin which can be integrated
into winmanager.

7. For the purpose of demonstrating the first feature, in this new version,
winmanager also ships with another lightweight plugin called
tagsexplorer.vim. This is an elementary plugin which shows the tags in the
present directory in a fancy manner and offers some simple ways of
navigating between them.
To see this in action [File List] window and press CTRL-N.